+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableBooleanArray.kt
@@ -366,10 +366,11 @@ public value class ImmutableBooleanArray @PublishedApi internal constructor(
         /**
          * Returns an immutable array containing the elements that were added.
          */
-        @Suppress("UNCHECKED_CAST")
         public fun build(): ImmutableBooleanArray {
-            if (size == 0) return EMPTY
-
+            when (size) {
+                0 -> return EMPTY
+                values.size -> return ImmutableBooleanArray(values)
+            }
             val backingArray = BooleanArray(size)
             System.arraycopy(values, 0, backingArray, 0, size)
             return ImmutableBooleanArray(backingArray)