+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableByteArray.kt
@@ -361,10 +361,11 @@ public value class ImmutableByteArray @PublishedApi internal constructor(
         /**
          * Returns an immutable array containing the elements that were added.
          */
-        @Suppress("UNCHECKED_CAST")
         public fun build(): ImmutableByteArray {
-            if (size == 0) return EMPTY
-
+            when (size) {
+                0 -> return EMPTY
+                values.size -> return ImmutableByteArray(values)
+            }
             val backingArray = ByteArray(size)
             System.arraycopy(values, 0, backingArray, 0, size)
             return ImmutableByteArray(backingArray)