+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableDoubleArray.kt
@@ -366,10 +366,11 @@ public value class ImmutableDoubleArray @PublishedApi internal constructor(
         /**
          * Returns an immutable array containing the elements that were added.
          */
-        @Suppress("UNCHECKED_CAST")
         public fun build(): ImmutableDoubleArray {
-            if (size == 0) return EMPTY
-
+            when (size) {
+                0 -> return EMPTY
+                values.size -> return ImmutableDoubleArray(values)
+            }
             val backingArray = DoubleArray(size)
             System.arraycopy(values, 0, backingArray, 0, size)
             return ImmutableDoubleArray(backingArray)